Generalizing Transactions to have outputs either NEList or []
Ignore-this: bc0b108e4040c37f35bd31d05e81839d

Normally there is a requirement that transaction outputs be non-empty;
however when performing a signature with SIGHASH_NONE, the signature
is made on a modified transaction with an empty list.  This we need our
data structure to be able to support empty lists in some cases.

We use a type parameter to statically ensure the lists are non-empty in the
usual case (using the type synonym Tx).

darcs-hash:20110709234806-e9b40-939e7da5097fd8c0f930f0b3ef52a8acbdabb0a7