Make addBlock independent of IO
Ignore-this: 70989addc75852ee9ff145cf1d45ca8a

In order for addBlock to be able to be used in atomiclyModifyIORef it cannot depend on IO.
Unfortunately this means that addBlock cannot enforce that it is passed the current time.
I'm not sure which is the better way to write this.

darcs-hash:20111102045917-e9b40-1e48bb6c8dcda7b42b044ece1a41b24399d2d116