addBlock enforces currentTime
Ignore-this: 48f223a0fe05d0ef2f216bd76a9ddd55

Ah solution!

This signature for addBlock lets us enforce that only the current time is used for addBlock and still lets is use atomicallyModifyIORef.
It is true that one could delay calling atomicallyModifyIORef, but the result is conservative.  It will err on rejecting blocks rather than accepting blocks.

Before this change there was the potential for someone to call addBlock with a time far far in the future leading to accepting a block that should be rejected (for now).

darcs-hash:20111102230435-e9b40-babbcbf1d6b02eb2de452a90c36e4c984125d826