AddBlockResult type
Ignore-this: d5238bc9f1ea7d9648067923c71ad9

Creating a data structure for the AddBlock return type so that it can indicate either the block is an orphan or if the block is in error.
This is in perparation for creating an orphanage.

darcs-hash:20120212062024-e9b40-604d9d8ad467b0993f0b198056da09bcd51fcae0